fix issue #346: Fixed a problem with the localoscillator management during Noto receiver configuration. The component is now ported into a standard nomencalature inside the repository. Also the code has been structured as Medicina component.
