+++ Noto/Servers/NotoPyLocalOscillator/src/Makefile
@@ -84,7 +84,7 @@ PY_SCRIPTS_L       =
 PY_MODULES         = 
 PY_MODULES_L       =
 
-PY_PACKAGES        =LocalOscillatorImpl 
+PY_PACKAGES        = NotoLocalOscillatorImpl 
 PY_PACKAGES_L      =
 pppppp_MODULES	   =
