+++ Noto/Servers/NotoPyLocalOscillator/src/NotoLocalOscillatorImpl/CommandLineError.py
@@ -1,0 +1,5 @@
+class CommandLineError(Exception):
+	def __init__(self, value):
+		self.value = value
+	def __str__(self):
+		return repr(self.value)
